% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{admixture2popSimulationWrapper}
\alias{admixture2popSimulationWrapper}
\title{Sample 2 pop simulated with ms}
\usage{
admixture2popSimulationWrapper(dataDirectory, nbIndiv, nbLocus, Nm, d,
  k = 1.9, force = FALSE, OnlyX = FALSE, admixe = TRUE, write = TRUE,
  snpsRate = 0)
}
\description{
Sample 2 pop simulated with ms
}

