% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{admixture2popSimulationWrapperWithOutlier}
\alias{admixture2popSimulationWrapperWithOutlier}
\title{Sample 2 pop simulated with ms. With outlier loci}
\usage{
admixture2popSimulationWrapperWithOutlier(nbIndiv, nbLocus, prop1, prop2, Nm1,
  Nm2, k, d = 1, admixe = FALSE)
}
\description{
Sample 2 pop simulated with ms. With outlier loci
}

