% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahelperfunction.R
\name{concatenateData}
\alias{concatenateData}
\title{Take dataset generated by a sampler an concatenate}
\usage{
concatenateData(dataDirectory, data1, data2, force = FALSE)
}
\description{
Take dataset generated by a sampler an concatenate
}

