% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahelperfunction.R
\name{freqInPop}
\alias{freqInPop}
\title{Compute allele frequency for each SNIPs in each pop.
Returns a matrix of size  nbLocus x nbPop}
\usage{
freqInPop(data, pop, ploidy)
}
\description{
Compute allele frequency for each SNIPs in each pop.
Returns a matrix of size  nbLocus x nbPop
}

