% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{Fst2Pop}
\alias{Fst2Pop}
\title{Compute the Fst of haplotype between 2 pops.}
\usage{
Fst2Pop(x)
}
\arguments{
\item{x}{A vector of allele (0 or 1) value in two pops with same number of individual.}
}
\description{
Compute the Fst of haplotype between 2 pops.
}

