% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{FstOF}
\alias{FstOF}
\title{Compute the Fst of haplotype with OF formula}
\usage{
FstOF(freq, q)
}
\arguments{
\item{freq}{A matrix of size nbLocus x nbPop}

\item{q}{A vector or list of size nbPop which contains the normalized size of pop}
}
\description{
Compute the Fst of haplotype with OF formula
}

