% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{lfmm_logit_sample}
\alias{lfmm_logit_sample}
\title{Sample genotype and environmental gradient.}
\usage{
lfmm_logit_sample(n, L, K, outlier_prop = 0, gamma = 1, gamma_mu = 1,
  gamma_U = 1, gamma_V = 1, gamma_B = 1, mean_B_outlier = 0.5,
  cor_U1_X = 0.5)
}
\description{
We use lfmm gibs sampler model to generate logit(P_ij) then we generate binary data G_ij
}

