% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSNPfoldedspectrum}
\alias{plotSNPfoldedspectrum}
\title{Plot folded allele frequency spectrum}
\usage{
plotSNPfoldedspectrum(genotype)
}
\arguments{
\item{genotype}{A matrix of genotype of size nbIndiv x nbLocus}
}
\description{
Plot folded allele frequency spectrum
}

