% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{rmse_withBestPermutation.greedy}
\alias{rmse_withBestPermutation.greedy}
\title{Compute the root mean square error between matrix by permuting
matrix column such that we have the best rmse. Warning : Greedy algorithm !}
\usage{
rmse_withBestPermutation.greedy(Q1, Q2)
}
\description{
Compute the root mean square error between matrix by permuting
matrix column such that we have the best rmse. Warning : Greedy algorithm !
}

