% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist.R
\name{rmse_withBestPermutation}
\alias{rmse_withBestPermutation}
\title{Compute the root mean square error between matrix by permuting
matrix column such that we have the best rmse.}
\usage{
rmse_withBestPermutation(Q1, Q2)
}
\description{
Compute the root mean square error between matrix by permuting
matrix column such that we have the best rmse.
}

