% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNMFWrapper.R
\name{RunsNMF}
\alias{RunsNMF}
\title{An sNMF wrapper for a good integration to R}
\usage{
RunsNMF(genotype, K, ploidy, alpha = 10, max.iteration = 200,
  tolerance = 1e-05)
}
\description{
An sNMF wrapper for a good integration to R
}

