% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFdS.R
\name{SFdS.runOnTrueData}
\alias{SFdS.runOnTrueData}
\title{for SFdS paper}
\usage{
SFdS.runOnTrueData(K, rep, file.X, file.coord, save.in.file, sigma = NULL,
  lambda = 1, max.iteration = 50)
}
\description{
for SFdS paper
}

