% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{SNPfoldedspectrum}
\alias{SNPfoldedspectrum}
\title{The folded allele frequency spectrum}
\usage{
SNPfoldedspectrum(genotype)
}
\arguments{
\item{genotype}{A matrix of genotype of size nbIndiv x nbLocus}
}
\description{
The folded allele frequency spectrum
}

