% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{SNPspectrum}
\alias{SNPspectrum}
\title{The allele frequency spectrum}
\usage{
SNPspectrum(genotype, allele = 1)
}
\arguments{
\item{genotype}{A matrix of genotype of size nbIndiv x nbLocus}

\item{allel}{0 or 1 for allele which you want the spectrum}
}
\description{
The allele frequency spectrum
}

