% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmpg2015.R
\name{sortWinner}
\alias{sortWinner}
\title{sort}
\usage{
sortWinner(result, score = function(fdr, precision) {     return(2 * ((1 -
  fdr) * precision)/((1 - fdr) + precision)) }, rankFun = sum)
}
\description{
sort
}

