% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{trueTPandFDR}
\alias{trueTPandFDR}
\title{Fdr control with benjamini hochgerg. This function
compute the true TP rate and FDR rate.}
\usage{
trueTPandFDR(p.values, alphaValues, outliers)
}
\description{
Fdr control with benjamini hochgerg. This function
compute the true TP rate and FDR rate.
}

